+++ sdks/python/src/opik/evaluation/samplers/__init__.py
@@ -1,3 +1,4 @@
-from .dataset_sampler import BaseDatasetSampler, RandomDatasetSampler
+from .base_dataset_sampler import BaseDatasetSampler
+from .random_dataset_sampler import RandomDatasetSampler
 
 __all__ = ["BaseDatasetSampler", "RandomDatasetSampler"]
